modified caxe_util.hpp Thread class to have size_t instead of int in return type of run. Avoid annoying warnings on int->pointer casts in 64bit

+++ incl/anal.hpp
@@ -12,7 +12,7 @@ class Analyser : public Thread {
     ref<tsDeque<ptr<MFile> > > files;
     ref<const std::vector<std::string> > search_dir;
 
-    int run();
+    size_t run();
 
 public:
