+++ incl/imports.hpp
@@ -11,7 +11,7 @@ class Importer : public Thread {
     ref<std::deque<ptr<MFile> >::iterator> ite;
     ref<tsDeque<ptr<MFile> > > files;
 
-    int run();
+    size_t run();
 
 public:
